+++ biz-front/src/components/molecules/MainPage/Popup/PopupRegist/BasicData/DescriptInputComp.tsx
