+++ biz-front/src/components/molecules/MainPage/Popup/PopupRegist/BasicData/ImagesInputcomp.tsx
